+++ .devcontainer/devcontainer.json
@@ -4,7 +4,7 @@
   "customizations": {
     "codespaces": {
       "openFiles": [
-        "article10.py"
+        "app.py"
       ]
     },
     "vscode": {
@@ -17,7 +17,7 @@
   },
   "updateContentCommand": "[ -f packages.txt ] && sudo apt update && sudo apt upgrade -y && sudo xargs apt install -y <packages.txt; [ -f requirements.txt ] && pip3 install --user -r requirements.txt; pip3 install --user streamlit; echo '✅ Packages installed and Requirements met'",
   "postAttachCommand": {
-    "server": "streamlit run article10.py --server.enableCORS false --server.enableXsrfProtection false"
+    "server": "streamlit run app.py --server.enableCORS false --server.enableXsrfProtection false"
   },
   "portsAttributes": {
     "8501": {
